pyproject.toml
[project]
name = "hackathon-agent"
version = "0.1.0"
description = "DesignForge AI - Design Automation Agent"
authors = [
    {name = "DesignForge Team", email = "team@designforge.ai"},
]
dependencies = [
    "langchain>=1.1.0",
    "langgraph>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph-sdk>=0.2.12",
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "pillow>=10.0.0",
    "firecrawl-py>=1.0.0",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

